Reorganize core
