+++ cli/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "try-runtime-cli"
+version = "0.1.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2021"
+license = "Apache-2.0"
+homepage = "https://substrate.io"
+repository = "https://github.com/paritytech/try-runtime-cli/"
+description = "CLI tool for runtime testing and dry-running"
+
+[[bin]]
+name = "try-runtime-cli"
+path = "main.rs"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+clap = { version = "4.0.9", features = ["derive"] }
+env_logger = "0.10.0"
+parity-scale-codec = "3.2.2"
+tokio = { version = "1.25.0", features = ["full"] }
+
+# Substrate
+sc-executor = { version = "0.22.0" }
+sp-io = { version = "19.0.0" }
+sp-runtime = { version = "20.0.0" }
+
+subxt = { git = "https://github.com/paritytech/subxt/", rev = "1c5faf3" }
+
+try-runtime-core = { path = "../core", features = ["cli"] }
