+++ core/Cargo.toml
@@ -13,7 +13,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 async-trait = "0.1.57"
-clap = { version = "4.0.9", features = ["derive"] }
+clap = { version = "4.0.9", features = ["derive"], optional = true }
 hex = { version = "0.4.3", default-features = false }
 log = "0.4.17"
 parity-scale-codec = "3.2.2"
@@ -28,11 +28,15 @@ frame-try-runtime = { version = "0.24.0", features = ["try-runtime"] }
 sc-cli = { version = "0.26.0" }
 sc-executor = { version = "0.22.0" }
 
+sp-api = { version = "16.0.0" }
 sp-core = { version = "18.0.0" }
 sp-externalities = { version = "0.18.0" }
 sp-io = { version = "19.0.0" }
 sp-runtime = { version = "20.0.0" }
 sp-state-machine = { version = "0.24.0" }
-sp-version = { version = "18.0.0" }
-sp-api = { version = "16.0.0" }
 sp-weights = { version = "16.0.0" }
+
+[features]
+cli = [
+    "clap",
+]
